Switch fonts when Chrome OS locale changes.
This patch is just a start.  When locale changes through the
login screen UI, the fonts in ResourceBundle and
PlatformFontGtk are updated, and the password fields and
username labels are redrawn with the correct font.  In the
future, all widgets in the login screen would be redrawn
with the correct font.

When the locale changes upon login, the fonts are updated,
though not all widgets are redrawn correctly yet.

BUG=chromium-os:13450
TEST=manually

Review URL: http://codereview.chromium.org/6788004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@80579 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28e466cf8282a13c22ff64eb762b7adef8f68f11