Split WGL implementations of *GLContext in GLContextWGL and *GLSurfaceWGL.
Surfaces are independent of contexts in GL. To facilitate sharing of surfaces between processes, I have separated them from the notion of contexts because contexts cannot be shared between processes.

I started with EGL in r81512. This is the same thing for WGL.

GLContextEGL still has a pointer to a surface and still has some surface specific operations that just forward through to it. Once I have refactored all the GLContext implementations in this way, I will remove these pointers and the surface specific opertations.

There will not be "view" and "offscreen" GL contexts. Rather there will be a single context type for each backend which can be made current with a surface that directs output either to a view or offscreen surface.
Review URL: http://codereview.chromium.org/6864002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@81807 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3e1b635a1f31e1c736ebb5dad2e772b12ac7aac