Load large icon resource pak
This is a part of r82185 that was reverted. The change was reverted because it caused a performance regression.

In r82538 re-checked in the code to split the high-res icons into a separate resource pak.

Now this change adds code to load that resource pak for Mac builds.

Once this passes the perf bot I'll submit additional changes to actuall use the new high-res icons.

BUG=None
TEST=None
Review URL: http://codereview.chromium.org/6895002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@82584 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d85ed75ba5aba12cd0b4d7d7676e17c19bebf58d