Split CGL implementations of *GLContext into GLContextCGL and *GLSurfaceCGL.
Surfaces are independent of contexts in GL. To facilitate sharing of surfaces between processes, I have separated them from the notion of contexts because contexts cannot be shared between processes.

I started with EGL in r81512, WGL in r81807, OSMesa in r81998 and GLX in r84090. This is the same thing for CGL.

GLContextCGL still has a pointer to a surface and still has some surface specific operations that just forward through to it. Once I have refactored all the GLContext implementations in this way, I will remove these pointers and the surface specific opertations.

There will not be "view" and "offscreen" GL contexts. Rather there will be a single context type for each backend which can be made current with a surface that directs output either to a view or offscreen surface.

TEST=WebGL locally, try
BUG=none
Review URL: http://codereview.chromium.org/6933020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@84334 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f96c7c517f99b8ed7490b1883fec672fcf2c2f2c