Remove override on thread IO restrictions.
This is the removal of code to disable security restrictions on disk IO,
originally needed so that the GL bindings could be initialized. The GL
libraries now manage this in a separate thread with the necessary permissions.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6990019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@86186 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90fd6ab9ba3443b8c746e852341c8b8ffbf33e52