This change implements a first pass in the effort to remove the dependency of PlatformDevice within Chrome.  The Skia library now provides multiple back-ends for the SkDevice class, so PlatformDevice's inheritance of SkDevice, and the assumption of instances of PlatformDevice limits the use of these new back-ends.
A new set of helper functions is provided for the PlatformDevice entry points.  Upon construction of a PlatformDevice, a pointer to the interface is cached in the parent SkDevice's SkMetaData.  The new helper functions forward calls to the interface cached in the metadata.

BUG=NONE
TEST=NONE
Review URL: http://codereview.chromium.org/7019013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@86625 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e414727ea30e03a5d4582bdbba42d67327605273