Read int properties from the X server as long.
The documentation says that the values are returned as long,
even though the range of values is only 32-bit.  This makes
it safe to cast from 64-bit to 32-bit.

GetIntProperty isn't used anywhere and GetIntArrayProperty is
only used on ChromiumOS, so it never hit this bug before.

Review URL: http://codereview.chromium.org/7020010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@87489 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5b7173002665650d9c0478cbe1a157fed31f7f5