ui/base/models: Fix the API example code comment.
Allocating |root| on the stack and passing to |model|, also on the stack,
does _not_ work, because TreeNodeModel owns the root.
So when the dtor of TreeNodeModel is called, it attempts to free an invalid
pointer, which was already deleted.

BUG=None
TEST=None

R=sky@chromium.org

Review URL: http://codereview.chromium.org/7171001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@89355 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3bfce524f3546661935e8448bad35a5571d8e22