Push (not pull) window geometry for the compositor.
The previous implementation used to poll the environment for the window size every drawing cycle. On Linux this is an expensive operation because it requires a round trip to the X server. This CL pushes window size change to the compositor.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7212028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@91572 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9884f501247498d6687ea762afbbab13e4045c9