Add invalid rect tracking and repaint scheduling to Layer.
This adds a LayerDelegate interface that can be implemented by an object that wants to know when to repaint.

This is a different model to current layer updating, where code must call SetCanvas when it wishes to push an updated canvas to the layer's texture.

BUG=none
TEST=see unittest.
Review URL: http://codereview.chromium.org/7822002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99080 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00b8698c798ba1fbc6994fdb8c28eff2be37b365