Reuse ResourceBundle to detect locale availability.
IsLocaleAvailable reimplements the logic for finding the resource file
for a given locale and testing if it exists. Expose this in
ResourceBundle instead which contains the actual resource file loading
code and thus is the canonical source for how to create this path.

TEST=ResourceBundle.LocaleDataPakExists

Review URL: http://codereview.chromium.org/7831036

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99495 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 756b6f4395615e32f6cf5ceb53d394cf393056e9