Redesignate checkmark resource and add themed IDR_INPUT_GOOD image
The resources IDR_INPUT_ALERT and IDR_INPUT_GOOD are used in tandem.  This CL makes them both part of the theme resources.  Furthermore, the IDR_INPUT_GOOD image is used by the checkmark throbber as a checkmark.  It is simply renamed to be as such.

This work is done in anticipation of using IDR_INPUT_ALRT and IDR_INPUT_GOOD in tandem as part of changing the edit search engine dialog to Web UI.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/8044019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@102771 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3e20fb203bc10e7c7d03ce5550a20371fd38872