views: Make sure layer visibility is updated properly.
If a View without a layer is hidden, and one of its descendant views has a layer, then the layer also needs to be hidden. This also needs to be reversed when such a View is redisplayed.

BUG=none
TEST=ViewLayerTreesInSync, VisibilityChildLayers

Review URL: http://codereview.chromium.org/8216027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@105080 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 25ae9a18091e849f488630a9f29e7023e5add08f