Downgrade a CHECK to a NOTREACHED in ResourceBundle::LoadLocaleResources.
This is a case where we're failing to mmap the file, even though it exists
on disk.  For example, we can't read the file or it has a size of 0.

We don't need to CHECK because in chrome_browser_main.cc, we prompt the
user to reinstall if this happens.

BUG=100395

Review URL: http://codereview.chromium.org/8301017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106138 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb76114572fbb128d2d8c4806d6f71e0da4f2835