Improve omnibox accessibility on Windows.
This refactors and improves a bunch of views accessibility
code on Windows and has the result of enabling NVDA to
announce the text selection in the omnibox.

The AutocompleteAccessibility class is removed; its
functionality is rolled into NativeViewAccessibilityWin.

NativeViewAccessibilityWin adds IAccessible2 and
IAccessibleText interfaces, allowing it to directly expose
the caret and selection to compatible assistive technology.
In addition, this gives each accessible object an unique id.

Finally, the reference from a View to its
NativeViewAccessibilityWin is changed from a scoped_refptr
to a ScopedComPtr, because another process may still have a
reference to the accessible COM object when the View is
deleted.

BUG=53380
TEST=Manually test the omnibox with NVDA on Windows. It should announce text that's selected, including autocompletions.
Review URL: http://codereview.chromium.org/8391010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@107924 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4003e72945f0faa549d21c90558d0dce96056806